Remove inits, combine sharding file
