--- sharktank/sharktank/ops/shaping/__init__.py
@@ -1,27 +1,0 @@
-# Copyright 2025 Advanced Micro Devices, Inc.
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-from .reshape import reshape
-from .view import view
-from .flatten import flatten
-from .unflatten import unflatten
-from .transpose import transpose
-from .permute import permute
-from .squeeze import squeeze
-from .unsqueeze import unsqueeze
-from .expand import expand
-
-__all__ = [
-    "reshape",
-    "view",
-    "flatten",
-    "unflatten",
-    "transpose",
-    "permute",
-    "squeeze",
-    "unsqueeze",
-    "expand",
-]
