+++ sharktank/sharktank/ops/shaping/transpose.py
@@ -16,7 +16,7 @@ from sharktank.types import (
     BlockScaledLayout,
 )
 from sharktank.ops._registry import overridable
-from ..sharding_utils import wrap_override
+from ..sharding.utils import wrap_override
 
 
 @overridable(dispatch_args=(0,))
